Overhauled project structure to be more organized.
