+++ src/components/home/stylesheet/sponsor_carousel.css
