+++ src/components/sponsors/SponsorTier.js
