+++ src/components/sponsors/Tier1Logo.js
