+++ src/components/sponsors/Tier2Logo.js
